drivers: i2c: stm32: Factorize i2c_stm32_runtime_configure()
Factorize i2c_stm32_runtime_configure() between STM32 I2C non-RTIO
and RTIO drivers that implements the same sequence but regarding
SMBus support that is not yet available in the RTIO driver.

No functional changes.
